# contracts/openapi/forgeml.v1.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelApprovalListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ModelApprovalResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "ModelApprovalListResponse",
        "type": "object"
      },
      "ModelApprovalResponse": {
        "properties": {
          "comment": {
            "title": "Comment",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model_version_id": {
            "title": "Model Version Id",
            "type": "string"
          },
          "policy_snapshot": {
            "additionalProperties": true,
            "title": "Policy Snapshot",
            "type": "object"
          },
          "requested_by": {
            "title": "Requested By",
            "type": "string"
          },
          "reviewer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "model_version_id",
          "status",
          "requested_by",
          "reviewer_id",
          "comment",
          "policy_snapshot"
        ],
        "title": "ModelApprovalResponse",
        "type": "object"
      },
      "ModelLineageListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ModelLineageResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "ModelLineageListResponse",
        "type": "object"
      },
...
      "ModelVersionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ModelVersionResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "ModelVersionListResponse",
        "type": "object"
      },
      "ModelVersionResponse": {
        "properties": {
          "artifact_manifest_hash": {
            "title": "Artifact Manifest Hash",
            "type": "string"
          },
          "artifact_manifest_uri": {
            "title": "Artifact Manifest Uri",
            "type": "string"
          },
          "artifact_uri": {
            "title": "Artifact Uri",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "experiment_run_id": {
            "title": "Experiment Run Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Metrics",
            "type": "object"
          },
          "model_format": {
            "title": "Model Format",
            "type": "string"
          },
          "registered_model_id": {
            "title": "Registered Model Id",
            "type": "string"
          },
          "signature": {
            "additionalProperties": true,
            "title": "Signature",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "training_run_id": {
            "title": "Training Run Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "registered_model_id",
          "version",
          "training_run_id",
          "experiment_run_id",
          "artifact_uri",
          "artifact_manifest_uri",
          "artifact_manifest_hash",
          "model_format",
          "signature",
          "metrics",
          "status",
          "created_by"
        ],
        "title": "ModelVersionResponse",
        "type": "object"
      },
...
      "PromoteTrainingRunRequest": {
        "properties": {
          "model_format": {
            "maxLength": 64,
            "minLength": 2,
            "title": "Model Format",
            "type": "string"
          },
          "signature": {
            "additionalProperties": true,
            "title": "Signature",
            "type": "object"
          },
          "training_run_id": {
            "title": "Training Run Id",
            "type": "string"
          }
        },
        "required": [
          "training_run_id",
          "model_format",
          "signature"
        ],
        "title": "PromoteTrainingRunRequest",
        "type": "object"
      },
...
      "RegisterModelVersionRequest": {
        "properties": {
          "model_format": {
            "maxLength": 64,
            "minLength": 2,
            "title": "Model Format",
            "type": "string"
          },
          "signature": {
            "additionalProperties": true,
            "title": "Signature",
            "type": "object"
          },
          "training_run_id": {
            "title": "Training Run Id",
            "type": "string"
          }
        },
        "required": [
          "training_run_id",
          "model_format",
          "signature"
        ],
        "title": "RegisterModelVersionRequest",
        "type": "object"
      },
      "RegisteredModelListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RegisteredModelResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "RegisteredModelListResponse",
        "type": "object"
      },
      "RegisteredModelResponse": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "owner_user_id": {
            "title": "Owner User Id",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_type": {
            "title": "Task Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "organization_id",
          "project_id",
          "name",
          "slug",
          "description",
          "task_type",
          "owner_user_id",
          "status"
        ],
        "title": "RegisteredModelResponse",
        "type": "object"
      },
      "RequestModelApprovalRequest": {
        "properties": {
          "comment": {
            "default": "",
            "maxLength": 2000,
            "title": "Comment",
            "type": "string"
          }
        },
        "title": "RequestModelApprovalRequest",
        "type": "object"
      },
...
      "ReviewModelVersionRequest": {
        "properties": {
          "comment": {
            "default": "",
            "maxLength": 2000,
            "title": "Comment",
            "type": "string"
          },
          "status": {
            "pattern": "^(approved|rejected)$",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "ReviewModelVersionRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "ForgeML",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/model-versions/{version_id}": {
      "get": {
        "operationId": "get_model_version_api_v1_model_versions__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Model Version",
        "tags": [
          "model-registry"
        ]
      }
    },
    "/api/v1/model-versions/{version_id}/approval-request": {
      "post": {
        "operationId": "request_model_approval_api_v1_model_versions__version_id__approval_request_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestModelApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelApprovalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Request Model Approval",
        "tags": [
          "model-registry"
        ]
      }
    },
    "/api/v1/model-versions/{version_id}/approvals": {
      "get": {
        "operationId": "list_model_approvals_api_v1_model_versions__version_id__approvals_get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelApprovalListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Model Approvals",
        "tags": [
          "model-registry"
        ]
      }
    },
    "/api/v1/model-versions/{version_id}/lineage": {
      "get": {
        "operationId": "list_model_lineage_api_v1_model_versions__version_id__lineage_get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelLineageListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Model Lineage",
        "tags": [
          "model-registry"
        ]
      }
    },
    "/api/v1/model-versions/{version_id}/review": {
      "post": {
        "operationId": "review_model_version_api_v1_model_versions__version_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewModelVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelApprovalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review Model Version",
        "tags": [
          "model-registry"
        ]
      }
    },
    "/api/v1/models/{model_id}": {
      "get": {
        "operationId": "get_registered_model_api_v1_models__model_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisteredModelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Registered Model",
        "tags": [
          "model-registry"
        ]
      }
    },
    "/api/v1/models/{model_id}/versions": {
      "get": {
        "operationId": "list_model_versions_api_v1_models__model_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelVersionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Model Versions",
        "tags": [
          "model-registry"
        ]
      },
      "post": {
        "operationId": "register_model_version_api_v1_models__model_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterModelVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Register Model Version",
        "tags": [
          "model-registry"
        ]
      }
    },
    "/api/v1/models/{model_id}/versions/promote-training-run": {
      "post": {
        "operationId": "promote_training_run_to_model_version_api_v1_models__model_id__versions_promote_training_run_post",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoteTrainingRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Promote Training Run To Model Version",
        "tags": [
          "model-registry"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/models": {
      "get": {
        "operationId": "list_registered_models_api_v1_projects__project_id__models_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisteredModelListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Registered Models",
        "tags": [
          "model-registry"
        ]
      },
      "post": {
        "operationId": "create_registered_model_api_v1_projects__project_id__models_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRegisteredModelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisteredModelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Registered Model",
        "tags": [
          "model-registry"
        ]
      }
    },
...
  }
}